# query/docs/swagger.json
{
    "consumes": [
        "application/json"
    ],
    "produces": [
        "application/json"
    ],
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "API Server Classroom service",
        "title": "Classroom API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/c/classroom",
    "paths": {
...
        "/grades/delete_program/{idModule}/{idProgram}": {
            "delete": {
                "description": "Delete grade program in module, ROLS=[teacher]",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "grades",
                    "classroom",
                    "roles.teacher"
                ],
                "summary": "Delete grade program",
                "parameters": [
                    {
                        "type": "string",
                        "description": "MongoID",
                        "name": "idModule",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "MongoID",
                        "name": "idProgram",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/res.Response"
                        }
                    },
                    "401": {
                        "description": "Unauthorized role",
                        "schema": {
                            "$ref": "#/definitions/res.Response"
                        }
                    },
                    "404": {
                        "description": "No existe la programación de calificación",
                        "schema": {
                            "$ref": "#/definitions/res.Response"
                        }
                    },
                    "409": {
                        "description": "Esta programación está en uso en alguna calificación",
                        "schema": {
                            "$ref": "#/definitions/res.Response"
                        }
                    },
                    "503": {
                        "description": "Service Unavailable - NATS || DB Service Unavailable\"..",
                        "schema": {
                            "$ref": "#/definitions/res.Response"
                        }
                    }
                }
            }
        },
...
        "/grades/update_grade/{idModule}/{idGrade}": {
            "put": {
                "description": "Update grade in module to student, ROLS=[teacher]",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "grades",
                    "classroom",
                    "roles.teacher"
                ],
                "summary": "Update grade",
                "parameters": [
                    {
                        "type": "string",
                        "description": "MongoID",
                        "name": "idModule",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "MongoID",
                        "name": "idGrade",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Desc",
                        "name": "grade",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/forms.UpdateGradeForm"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/res.Response"
                        }
                    },
                    "400": {
                        "description": "Calificación inválida. Mín: %v. Máx: %v",
                        "schema": {
                            "$ref": "#/definitions/res.Response"
                        }
                    },
                    "401": {
                        "description": "Unauthorized role",
                        "schema": {
                            "$ref": "#/definitions/res.Response"
                        }
                    },
                    "404": {
                        "description": "No existe la programación de calificación",
                        "schema": {
                            "$ref": "#/definitions/res.Response"
                        }
                    },
                    "409": {
                        "description": "Esta calificación no pertenece al módulo",
                        "schema": {
                            "$ref": "#/definitions/res.Response"
                        }
                    },
                    "503": {
                        "description": "Service Unavailable - NATS || DB Service Unavailable\"..",
                        "schema": {
                            "$ref": "#/definitions/res.Response"
                        }
                    }
                }
            }
        },
        "/grades/upload_grade/{idModule}/{idStudent}": {
            "post": {
                "description": "Upload grade in module to student, ROLS=[teacher]",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "grades",
                    "classroom",
                    "roles.teacher"
                ],
                "summary": "Upload grade",
                "parameters": [
                    {
                        "type": "string",
                        "description": "MongoID",
                        "name": "idModule",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "MongoID",
                        "name": "idStudent",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Desc",
                        "name": "grade",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/forms.GradeForm"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/res.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "body": {
                                            "$ref": "#/definitions/smaps.IdInsertedMap"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Calificación inválida. Mín: %v. Máx: %v",
                        "schema": {
                            "$ref": "#/definitions/res.Response"
                        }
                    },
                    "401": {
                        "description": "Unauthorized role",
                        "schema": {
                            "$ref": "#/definitions/res.Response"
                        }
                    },
                    "409": {
                        "description": "No se puede agregar una calificación ya subida",
                        "schema": {
                            "$ref": "#/definitions/res.Response"
                        }
                    },
                    "503": {
                        "description": "Service Unavailable - NATS || DB Service Unavailable",
                        "schema": {
                            "$ref": "#/definitions/res.Response"
                        }
                    }
                }
            }
        },
        "/grades/upload_program/{idModule}": {
            "post": {
                "description": "Upload program grade in module, ROLS=[teacher]",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "grades",
                    "classroom",
                    "roles.teacher"
                ],
                "summary": "Upload program grade",
                "parameters": [
                    {
                        "type": "string",
                        "description": "MongoID",
                        "name": "idModule",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Desc",
                        "name": "programGrade",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/forms.GradeProgramForm"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/res.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "body": {
                                            "$ref": "#/definitions/smaps.IdInsertedMap"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "El porcentaje sumatorio de las calificaciones acumulativas debe ser exactamente 100 por cierto",
                        "schema": {
                            "$ref": "#/definitions/res.Response"
                        }
                    },
                    "401": {
                        "description": "Unauthorized role",
                        "schema": {
                            "$ref": "#/definitions/res.Response"
                        }
                    },
                    "409": {
                        "description": "Esta calificación ya está programada",
                        "schema": {
                            "$ref": "#/definitions/res.Response"
                        }
                    },
                    "503": {
                        "description": "Service Unavailable - NATS || DB Service Unavailable",
                        "schema": {
                            "$ref": "#/definitions/res.Response"
                        }
                    }
                }
            }
        },
...
        "/modules/new_sub_section/{idModule}": {
            "post": {
                "description": "New sub section to module",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "modules",
                    "classroom",
                    "roles.teacher"
                ],
                "summary": "New sub section",
                "parameters": [
                    {
                        "type": "string",
                        "description": "MongoID",
                        "name": "idModule",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/res.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "body": {
                                            "$ref": "#/definitions/smaps.InsertedIdMap"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad path params",
                        "schema": {
                            "$ref": "#/definitions/res.Response"
                        }
                    },
                    "401": {
                        "description": "Unauthorized role",
                        "schema": {
                            "$ref": "#/definitions/res.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "forms.AcumulativeForm": {
            "type": "object",
            "required": [
                "number",
                "percentage"
            ],
            "properties": {
                "number": {
                    "type": "integer",
                    "minimum": 1,
                    "example": 1
                },
                "percentage": {
                    "type": "number",
                    "example": 45
                }
            }
        },
...
        "forms.GradeForm": {
            "type": "object",
            "required": [
                "grade",
                "program"
            ],
            "properties": {
                "acumulative": {
                    "type": "string",
                    "example": "637ab12ae976057567a4d67d"
                },
                "grade": {
                    "type": "number",
                    "example": 70
                },
                "program": {
                    "type": "string",
                    "example": "637ab12ae976057567a4d67d"
                }
            }
        },
        "forms.GradeProgramForm": {
            "type": "object",
            "required": [
                "is_acumulative",
                "number",
                "percentage"
            ],
            "properties": {
                "acumulative": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/forms.AcumulativeForm"
                    }
                },
                "is_acumulative": {
                    "type": "boolean"
                },
                "number": {
                    "type": "integer",
                    "maximum": 30,
                    "minimum": 1,
                    "example": 1
                },
                "percentage": {
                    "type": "number",
                    "example": 25
                }
            }
        },
...
        "forms.UpdateGradeForm": {
            "type": "object",
            "required": [
                "grade"
            ],
            "properties": {
                "grade": {
                    "type": "number",
                    "example": 55
                }
            }
        },
...
        "res.Response": {
            "type": "object",
            "properties": {
                "body": {
                    "type": "object",
                    "additionalProperties": true
                },
                "message": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
...
        "smaps.IdInsertedMap": {
            "type": "object",
            "properties": {
                "_id": {
                    "type": "string"
                }
            }
        },
        "smaps.InsertedIdMap": {
            "type": "object",
            "properties": {
                "inserted_id": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "description": "BearerJWTToken in Authorization Header",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    },
...
}